move all context types to same file
